(fix) deleted unused method and fixed count data type to int

+++ apiserver/service/src/main/java/com/zipline/service/statics/StaticsService.java
@@ -1,8 +1,8 @@
 package com.zipline.service.statics;
 
 public interface StaticsService {
-  long getRecentContractCount(Long userId);
-  long getOngoingContractCount(Long userId);
-  long getCompletedContractCount(Long userId);
-  long getRecentCustomerCount(Long userId);
+  int getRecentContractCount(Long userId);
+  int getOngoingContractCount(Long userId);
+  int getCompletedContractCount(Long userId);
+  int getRecentCustomerCount(Long userId);
 }
